Refatorado o método de atualização e remoção das entidades Doacao e Usuario

+++ src/main/java/br/com/pawsoncloud/servicos/impl/DoacaoServicoImpl.java
@@ -2,6 +2,7 @@ package br.com.pawsoncloud.servicos.impl;
 
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.dao.EmptyResultDataAccessException;
+import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.stereotype.Service;
 
 import br.com.pawsoncloud.dto.DoacaoDto;
@@ -10,6 +11,7 @@ import br.com.pawsoncloud.entidades.Doacao;
 import br.com.pawsoncloud.repositorios.DoacaoRepositorio;
 import br.com.pawsoncloud.servicos.DoacaoServico;
 import br.com.pawsoncloud.servicos.conversor.DadosDoacao;
+import br.com.pawsoncloud.servicos.conversor.UsuarioLogado;
 import br.com.pawsoncloud.servicos.excecoes.DataBaseException;
 import br.com.pawsoncloud.servicos.excecoes.ObjectNotFoundException;
 import jakarta.persistence.EntityNotFoundException;
@@ -31,8 +33,12 @@ public class DoacaoServicoImpl implements DoacaoServico {
     public Doacao update(Long id, DoacaoUpdateDto doacaoDto) {
         try {
             Doacao doacao = repositorio.getReferenceById(id);
-            DadosDoacao.getDoacaoAtualizada(doacao, doacaoDto);
-            return repositorio.save(doacao);
+            if (doacao.getDoador().equals(UsuarioLogado.getUsuario())) {
+                DadosDoacao.getDoacaoAtualizada(doacao, doacaoDto);
+                return repositorio.save(doacao);
+            } else {
+                throw new BadCredentialsException("Não autorizado");
+            }
         } catch (EntityNotFoundException e) {
             throw new ObjectNotFoundException("Doação não encontrada. Id inválido: " + id);
         }
@@ -41,7 +47,12 @@ public class DoacaoServicoImpl implements DoacaoServico {
     @Override
     public void delete(Long id) {
         try {
-            repositorio.deleteById(id);
+            Doacao doacao = repositorio.getReferenceById(id);
+            if (doacao.getDoador().equals(UsuarioLogado.getUsuario())) {
+                repositorio.deleteById(id);
+            } else {
+                throw new BadCredentialsException("Não autorizado");
+            }
         } catch (EmptyResultDataAccessException e) {
             throw new ObjectNotFoundException("Doação não encontrada. Id inválido: " + id);
         } catch (DataIntegrityViolationException e) {
